Move spinner to its own folder
